Update operators (#2927)
* update list plugins to obey enable/disable

* add wrapper method for backwards compat
